+++ NEWS.md
@@ -9,6 +9,7 @@
 * Fixed incorrect selection after feature table (or other interactive tables) have been manually re-ordered (reported by Thanh Wang)
 * `groups()` and `as.data.table()` methods for `featureGroups`: optionally consider feature areas instead of peak intensities.
 * `plotSilhouettes()` method for `compoundsCluster`
+* Added `rGroups` argument to subset operator for `featureGroups` to subset by replicate groups (equivalent to `rGroups` argument to `filter()`).
 
 ## May 2019